requirements.txt
cx_Oracle==8.3.0
gspread==5.7.0
oauth2client==4.1.3
pandas==2.1.0
python-dotenv==1.0.0
loguru==0.7.0